+++ src/lib/scriptengine/rsscript.rs
@@ -45,6 +45,7 @@ fn install_crate(provider: &ScriptRunner) {
                 min_version: None,
                 version: None,
                 install_command: None,
+                force: None,
             };
 
             // due to fornwall/rust-script/issues/42
@@ -60,6 +61,7 @@ fn install_crate(provider: &ScriptRunner) {
             true,
             &None,
             &None,
+            &None,
         ),
         ScriptRunner::CargoPlay => cargo_plugin_installer::install_crate(
             &None,
@@ -67,6 +69,7 @@ fn install_crate(provider: &ScriptRunner) {
             "cargo-play",
             &None,
             true,
+            &None,
             &None,
             &None,
         ),