+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## CHANGELOG
 
+### v0.35.16
+
+* Enhancement: Create task definition on runtime and invoke it via plugin #677
+* Enhancement: Enable to remove force flag for custom install commands #674
+* Maintenance: Update dependencies #676
+
 ### v0.35.15 (2022-07-21)
 
 * Fix: upgrade cliparser to support windows paths